auth-service/pyproject.toml
[project]
name = "service-base-auth-fastapi"
version = "0.1.0"
description = "A FastAPI-based authentication service with PostgreSQL, using FastAPI Users, SQLAlchemy, and Alembic."
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "aioredis>=2.0.1",
    "aioshutil>=1.5",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "backports-cached-property>=1.0.2",
    "boto3>=1.38.3",
    "colorama>=0.4.6",
    "cryptography>=44.0.2",
    "cython>=3.0.12",
    "databases>=0.9.0",
    "editdistance>=0.8.1",
    "factory-boy>=3.3.3",
    "fastapi>=0.115.12",
    "fastapi-mail>=1.4.2",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "fastapi-users[oauth,sqlalchemy]>=14.0.1",
    "hashids>=1.3.1",
    "httpx==0.28.1",
    "marshmallow>=4.0.0",
    "mypy>=1.15.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "protobuf>=3.20.2",
    "psycopg2>=2.9.10",
    "py3langid==0.3.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pytest-factoryboy>=2.7.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "ruff>=0.11.7",
    "sqlalchemy>=2.0.40",
    "stripe>=12.1.0",
    "tqdm>=4.67.1",
    "ulid-py>=1.1.0",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
    "sendria>=2.2.2",
]

[tool.setuptools]
packages = ["app", "alembic", "tests", "scripts", "docker"]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "black>=25.1.0",
    "boto3-stubs>=1.38.8",
    "faker>=37.1.0",
    "freezegun>=1.5.1",
    "ipykernel==6.30.1",
    "jupyter>=1.1.1",
    "pylint==3.3.8",
    "pytest==7.2.2",
    "pytest-asyncio==0.21.1",
    "pytest-cov>=6.1.1",
    "pytest-dotenv>=0.5.2",
    "pytest-sugar==0.9.6",
    "pytest-xdist==3.8.0",
]


[tool.ruff]
line-length = 88  # または好きな長さ (PEP8は79, Blackは88)
target-version = "py310"  # プロジェクトに合わせて (例: Python 3.10なら "py310")
fix = true
unsafe-fixes = false


[tool.ruff.format]
quote-style = "double"  # 文字列はダブルクオート（好みで'single'でもOK）
docstring-code-format = true

[tool.mypy]
files = ["app", "alembic", "tests", "scripts"]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["F401", "F403"]
"tests/v1/conftest.py" = ["F401"]
"alembic/env.py" = ["F401", "F403"]
